Moves shell option helpers to lib.process_helpers
